+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Next release
 
 - ci: add coveralls report
+- fix: fixed the starting block arg with an ignore_block_order argument
 - docs: fixed Docker Compose instructions
 - fix: removed unused dependencies with udeps and machete
 - feat: add devnet via `--devnet` cli argument